requirements.txt
crewai
gradio
markdown_pdf
openai
python-dotenv
crewai_tools 
agentops
fastapi 
uvicorn
requests